requirements.txt
paho-mqtt
dash
plotly
pandas
numpy
scikit-learn
dash-daq
dash-bootstrap-components
rrcf
